backend/participation-service/requirements.txt
Flask~=2.2.3
Flask-SQLAlchemy~=3.0.3
Flask-Migrate~=4.0.4
Flask-APScheduler~=1.12.4
psycopg2-binary~=2.9.5
requests~=2.28.2